patches: Fix runtime issues
The first patch disables the setting of the process title in Nginx. This
is done because of the way Nginx accomplishes this on Linux -it can be
found as a comment in ngx_setproctitle.c- which would not work in Unikraft.

The second patch disables semaphore sharing between processes in Nginx.

Signed-off-by: Mihai Pogonaru <pogonarumihai@gmail.com>
Reviewed-by: Costin Lupu <costin.lupu@cs.pub.ro>