Config.uk: Imply LIBUKMMAP
Change `select LIBUKMMAP` to `imply LIBUKMMAP`. This is because
`LIBUKMMAP` may not be required for the Nginx build (and use
`LIBPOSIX_MMAP` instead).

With `select` in place you cannot disable `LIBUKMMAP`
resulting in a dependency erorr when enabling `LIBPOSIX_MMAP`.

With `imply` as a soft select, `LIBUKMMAP` can be disabled, removing the
dependency error.

Co-authored-by: Marc Rittinghaus <marc@unikraft.io>
Signed-off-by: Marc Rittinghaus <marc@unikraft.io>
Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>